Replace StandardDialog with Melt StandardDialog
